+++ packages/mcp-lexicon/src/coding/lsp/mod.rs
@@ -52,16 +52,37 @@
 //! ```
 
 pub mod client;
+pub mod common;
 pub mod diagnostics;
 pub mod error;
+pub mod lsp_diagnostics;
+pub mod lsp_find_references;
+pub mod lsp_goto_definition;
+pub mod lsp_hover;
+pub mod lsp_workspace_symbol;
 pub mod transport;
 
 pub use client::{
     LspClient, NotificationReceiver, NotificationSender, ServerNotification, path_to_uri,
 };
+pub use common::{LocationResult, parse_line, symbol_kind_to_string, uri_to_path};
 pub use diagnostics::{
     DiagnosticCounts, FormattedDiagnostic, Severity, count_by_severity, filter_by_severity,
     format_diagnostics,
 };
 pub use error::{LspError, Result};
+pub use lsp_diagnostics::{
+    DiagnosticsSummary, LspDiagnostic, LspDiagnosticsInput, LspDiagnosticsOutput,
+    execute_lsp_diagnostics,
+};
+pub use lsp_find_references::{
+    LspFindReferencesInput, LspFindReferencesOutput, execute_lsp_find_references,
+};
+pub use lsp_goto_definition::{
+    LspGotoDefinitionInput, LspGotoDefinitionOutput, execute_lsp_goto_definition,
+};
+pub use lsp_hover::{LspHoverInput, LspHoverOutput, execute_lsp_hover};
+pub use lsp_workspace_symbol::{
+    LspWorkspaceSymbolInput, LspWorkspaceSymbolOutput, SymbolResult, execute_lsp_workspace_symbol,
+};
 pub use transport::{ClientNotification, LanguageId};
